src/requirements.txt
requests==2.31.0
Flask==3.0.0
Pillow==9.2.0
pandas==2.1.4
jinja2==3.1.2
python-docx==0.8.11
